@@ -7,7 +7,7 @@
             slidesToScroll: 1,
             dots: false,
             arrows: false,
-            autoplay: true,
+            autoplay: false,
             autoplaySpeed: 50
         });
     });
